Big update to expose object -> world transform.
Spire now requires some common uniforms to be in place in order for
world transformations to be handled efficiently. The parent program will
be able to change the names of these common uniforms in the future. At
the very least, spire needs to know where to find the camera and
projection transforms, and where to place the appropriately transformed
world coordinates.