Refactoring: Converting StuInterface thread functions.
All functions are being added to the interface implementaiton class.
Opted for this route because of the clear line of demarcation between
functions operating on the Spire thread and functions that do not.
Functions in the Interface class are all thread safe, with exceptions
noted. All InterfaceImplementation functions are running on the Spire
thread.