Reverse light vector. See further explanation below.
"The light vector l is usually defined pointing in a direction opposite
to the direction the light is traveling." - Real-Time Rendering, page
101.

I agree with this approach. I was performing a negation in the shaders
in order to orient the light vector in the direction light is
travelling; since defining the light vector as opposite the direction
light is traveling is common in graphics, I adopt the notion here.